+++ src/main/scala/ScalatraBootstrap.scala
@@ -31,7 +31,7 @@ class ScalatraBootstrap extends LifeCycle with StrictLogging {
       if (!configFile.canRead) {
         throw new ServiceConfigurationError("Could not read configuration file " + configFile)
       }
-      ConfigFactory.parseFile(configFile)
+      ConfigFactory.parseFile(configFile).resolve()
     }
 
     implicit val setup = new Setup(config, new Emailer(config.getConfig("email")))