+++ src/main/scala/org/mmisw/orr/ont/app/SelfHostedOntController.scala
@@ -33,7 +33,9 @@ class SelfHostedOntController(implicit setup: Setup, ontService: OntService) ext
     }
 
     if (!portalDispatch(pathInfo, reqFormat)) {
-      resolve(pathInfo.substring(1), reqFormat)
+      // skip leading slash if any
+      val noSlash = if (pathInfo.length() > 1) pathInfo.substring(1) else pathInfo
+      resolve(noSlash, reqFormat)
     }
   }
 
@@ -105,7 +107,7 @@ class SelfHostedOntController(implicit setup: Setup, ontService: OntService) ext
               case exc: AnyRef =>
                 logger.info(s"EXC in selfResolve: $exc")
                 // TODO dispatch some synthetic response as in previous Ont
-                error(500, s"TODO: generate summary for organization '$xyz'")
+                error500(s"TODO: generate summary for organization '$xyz'")
             }
         }
       case None =>
@@ -114,7 +116,7 @@ class SelfHostedOntController(implicit setup: Setup, ontService: OntService) ext
             user.ontUri match {
               case Some(ontUri) => redirect(ontUri)
               case None =>
-                error(500, s"TODO: generate summary for user '$xyz'")
+                error500(s"TODO: generate summary for user '$xyz'")
             }
           case None => error(404, s"No organization or user by given name: '$xyz'")
         }
@@ -158,7 +160,7 @@ class SelfHostedOntController(implicit setup: Setup, ontService: OntService) ext
     Try(ontService.resolveOntology(uri, versionOpt)) match {
       case Success(res)         => res
       case Failure(exc: NoSuch) => error(404, exc.details)
-      case Failure(exc)         => error(500, exc.getMessage)
+      case Failure(exc)         => error500(exc)
     }
   }
 
@@ -166,7 +168,7 @@ class SelfHostedOntController(implicit setup: Setup, ontService: OntService) ext
     Try(ontService.getOntologyFile(uri, version, reqFormat)) match {
       case Success(res)                   => res
       case Failure(exc: NoSuchOntFormat)  => error(406, exc.details)
-      case Failure(exc) => error(500, exc.getMessage)
+      case Failure(exc)                   => error500(exc)
     }
   }